+++ src/db/Database.h
@@ -65,14 +65,14 @@ public:
 // 	virtual bool removeProductsRecord(const QModelIndex & idx, bool askForConfirmation = true);
 	virtual bool removeProductsRecord(const QModelIndexList & idxl, bool askForConfirmation = true);
 
-	virtual bool addBatchRecord(int pid, const QString& spec, const QString& book, const QString& reg, const QString& expiry, float qty, float used, const QString& unit, const QString& price, const QString& invoice, const QString& notes);
-	virtual bool updateBatchRecord(int bid, int pid, const QString& spec, const QString& book, const QString& reg, const QString& expiry, float qty, float used, const QString& unit, const QString& price, const QString& invoice, const QString& notes);
+	virtual bool addBatchRecord(int pid, const QString& spec, const QString& book, const QString& reg, const QString& expiry, double qty, double used, const QString& unit, const QString& price, const QString& invoice, const QString& notes);
+	virtual bool updateBatchRecord(int bid, int pid, const QString& spec, const QString& book, const QString& reg, const QString& expiry, double qty, double used, const QString& unit, const QString& price, const QString& invoice, const QString& notes);
 // 	virtual bool removeBatchRecord(int recordid, bool askForConfirmation = true);
 // 	virtual bool removeBatchRecord(const QModelIndex & idx, bool askForConfirmation = true);
 	virtual bool removeBatchRecord(const QModelIndexList & idxl, bool askForConfirmation = true);
 
-	virtual bool addDistributorRecord(int, float, const QString&, const QString&, const QString&, const QString&, DistributorTableModel::Reasons);
-	virtual bool updateDistributorRecord(int, int, float, const QString&, const QString&, const QString&, const QString&, DistributorTableModel::Reasons);
+	virtual bool addDistributorRecord(int, double, const QString&, const QString&, const QString&, const QString&, DistributorTableModel::Reasons);
+	virtual bool updateDistributorRecord(int, int, double, const QString&, const QString&, const QString&, const QString&, DistributorTableModel::Reasons);
 	virtual bool removeDistributorRecord(int recordid, bool askForConfirmation = true, bool submitBatches = true);
 // 	virtual bool removeDistributorRecord(const QModelIndex & idx, bool askForConfirmation = true, bool submitBatches = true);
 	virtual bool removeDistributorRecord(const QModelIndexList & idxl, bool askForConfirmation = true, bool submitBatches = true);
